+++ CHANGELOG.md
@@ -61,6 +61,8 @@ Main (unreleased)
 
 - Support Scrape Protocol specification in CRDS for `prometheus.operator.*` components. (@dehaansa) 
 
+- Fix `otelcol.exporter.splunkhec` arguments missing documented `otel_attrs_to_hec_metadata` block. (@dehaansa)
+
 v1.11.2
 -----------------